feat(webui): move events into a settings tab
The events view is a debugging tool that no longer warrants top-nav real
estate. Move it under a new Events tab in Settings alongside the routing
rules, drop the top-nav link, and redirect /events to the new tab so old
links continue to work.

Closes #693
