+++ webui/src/routes/events.index.tsx
@@ -1,147 +1,7 @@
-import { useState } from 'react'
-import { createFileRoute, Link } from '@tanstack/react-router'
-import { useQuery } from '@connectrpc/connect-query'
-import { listEvents } from '@/gen/xagent/v1/xagent-XAgentService_connectquery'
-import type { Event } from '@/gen/xagent/v1/xagent_pb'
-import { timestampDate } from '@bufbuild/protobuf/wkt'
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from '@/components/ui/table'
-import { RelativeTime } from '@/components/relative-time'
-import { Button } from '@/components/ui/button'
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select'
-import { Plus } from 'lucide-react'
-import { useOrgId } from '@/hooks/use-org-id'
+import { createFileRoute, redirect } from '@tanstack/react-router'
 
 export const Route = createFileRoute('/events/')({
-  component: EventsPage,
+  beforeLoad: () => {
+    throw redirect({ to: '/settings', search: { tab: 'events' }, replace: true })
+  },
 })
-
-function EventsPage() {
-  const orgId = useOrgId()
-  const [limit, setLimit] = useState(25)
-  const { data, isLoading, error } = useQuery(
-    listEvents,
-    { limit },
-    {
-      refetchInterval: 60000,
-    },
-  )
-
-  if (isLoading) {
-    return (
-      <div className="flex min-h-screen items-center justify-center">
-        <div className="text-muted-foreground">Loading events...</div>
-      </div>
-    )
-  }
-
-  if (error) {
-    return (
-      <div className="flex min-h-screen items-center justify-center">
-        <div className="text-destructive">Error: {error.message}</div>
-      </div>
-    )
-  }
-
-  const events = data?.events ?? []
-
-  return (
-    <div className="container mx-auto py-8 px-4">
-      <div className="flex flex-wrap items-center justify-between gap-4 mb-6">
-        <h1 className="text-2xl font-bold">Recent Events</h1>
-        <div className="flex flex-wrap items-center gap-4">
-          <div className="flex items-center gap-2">
-            <span className="text-sm text-muted-foreground">Show</span>
-            <Select value={String(limit)} onValueChange={(value) => setLimit(Number(value))}>
-              <SelectTrigger className="w-20">
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="25">25</SelectItem>
-                <SelectItem value="50">50</SelectItem>
-                <SelectItem value="75">75</SelectItem>
-                <SelectItem value="100">100</SelectItem>
-              </SelectContent>
-            </Select>
-          </div>
-          <Link to="/events/new" search={{ org: orgId }}>
-            <Button>
-              <Plus className="h-4 w-4" />
-              Event
-            </Button>
-          </Link>
-        </div>
-      </div>
-      {events.length === 0 ? (
-        <div className="text-muted-foreground text-center py-8">No events found</div>
-      ) : (
-        <Table>
-          <TableHeader>
-            <TableRow>
-              <TableHead className="hidden md:table-cell">ID</TableHead>
-              <TableHead>Description</TableHead>
-              <TableHead>Data</TableHead>
-              <TableHead className="hidden md:table-cell">Created</TableHead>
-            </TableRow>
-          </TableHeader>
-          <TableBody>
-            {events.map((event) => (
-              <EventRow key={String(event.id)} event={event} />
-            ))}
-          </TableBody>
-        </Table>
-      )}
-    </div>
-  )
-}
-
-function EventRow({ event }: { event: Event }) {
-  const orgId = useOrgId()
-  const dataContent = event.data || '-'
-  const truncatedData = dataContent.length > 100 ? dataContent.slice(0, 100) + '...' : dataContent
-
-  return (
-    <TableRow>
-      <TableCell className="hidden md:table-cell">{String(event.id)}</TableCell>
-      <TableCell>
-        <Link
-          to="/events/$id"
-          search={{ org: orgId }}
-          params={{ id: String(event.id) }}
-          className="text-primary hover:underline"
-        >
-          {event.description || '-'}
-        </Link>
-      </TableCell>
-      <TableCell className="max-w-xs truncate">
-        {event.url ? (
-          <a
-            href={event.url}
-            target="_blank"
-            rel="noopener noreferrer"
-            className="text-primary hover:underline"
-          >
-            {truncatedData}
-          </a>
-        ) : (
-          truncatedData
-        )}
-      </TableCell>
-      <TableCell className="hidden md:table-cell text-muted-foreground">
-        {event.createdAt ? <RelativeTime date={timestampDate(event.createdAt)} /> : '-'}
-      </TableCell>
-    </TableRow>
-  )
-}
